backend/requirements.txt
Django>=5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.5
django-filter>=24.3
drf-nested-routers>=0.93
drf-spectacular>=0.27
psycopg[binary]>=3.2
gunicorn>=22.0
pillow>=10.4
python-decouple>=3.8
